test: abstract bucket utils
